updates backend with eslint and tsconfig
