+++ backend/Dockerfile
@@ -1,10 +1,9 @@
 FROM        node:latest
-ENV         NODE_ENV=production
 ENV         PORT=5000
 COPY        . /src
 WORKDIR     /src
 RUN         npm install
 EXPOSE      5000
 COPY . .
-RUN tsc
-ENTRYPOINT  ["node", "start"]
+RUN         npm run build
+ENTRYPOINT  ["node", "dist/index.js"]
