--- backend/jest.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  preset: "ts-jest",
-  testEnvironment: "node",
-  collectCoverageFrom: ["src/*/**/*.ts"],
-  testPathIgnorePatterns: ["/node_modules/", "tsbuild"],
-  modulePathIgnorePatterns: ["<rootDir>/.*/__mocks__"],
-  roots: ["src"],
-};
