--- backend/src/routes/routes.ts
@@ -1,9 +1,0 @@
-import express, { Request, Response } from 'express'
-
-const router = express.Router();
-
-router.get('/', (req: Request, res: Response) => {
-    res.send("Hello world!");
-})
-
-export default router;
