--- backend/tslint.json
@@ -1,15 +1,0 @@
-{
-    "defaultSeverity": "error",
-    "extends": [
-        "tslint:recommended"
-    ],
-    "jsRules": {},
-    "rules": {
-        "no-console": false,
-        "no-string-literal": false,
-        "comment-format": false,
-        "jsdoc-format": false,
-        "object-literal-shorthand": false
-    },
-    "rulesDirectory": []
-}
